Split DnaDef and DnaFile into own modules
